Fix component setup from addon to app
New setup allows using the same template for multiple component names.
This fixes the `{{new-version-notifier}}` not working previously, since
it didn't have a template (it needed a template with it's own name). This
fix allows one template to be used for both.

Also added a simple test in the dummy app with a VERSION.txt