feat: Always create the version file
BREAKING CHANGE: The version file is always created, you must opt out now instead of opt in, see the README

The way to opt out of this behavior is to include a VERSION.txt in the
app, otherwise the file will be created automatically. When a
VERSION.txt file exists the treeForPublic build step won't overwrite it.

Co-authored-by: Ilya Radchenko <knownasilya@gmail.com>